requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioschedule==0.5.2
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-apscheduler==0.7.0
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-ckeditor5==0.0.4
django-cors-headers==4.6.0
django-cron==0.6.0
django-cuser==2017.3.16
django-js-asset==3.0.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-prometheus==2.3.1
django-redis==5.4.0
django-timezone-field==7.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
magic-filter==1.0.12
modeltranslation==0.25
multidict==6.0.5
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.10.0
pytest==8.3.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
users==1.0.dev0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.7